requirements.txt
Pillow==10.1.0
streamlit==1.38.0